requirements.txt
discord
python-dotenv
pyautogui
sounddevice
numpy
scipy